% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxes.R
\name{carouselBox}
\alias{carouselBox}
\alias{carouselItem}
\alias{carouselSets}
\title{Create a carousel box for the main body of a dashboard.}
\usage{
carouselBox(..., title = NULL, footer = NULL, status = NULL,
  solidHeader = FALSE, background = NULL, width = 6, height = NULL,
  collapsible = FALSE, collapsed = FALSE)

carouselSets(..., id = NULL, class = NULL, itemStartWith = 1L,
  .list = NULL)

carouselItem(..., caption = NULL, .list = NULL)
}
\arguments{
\item{...}{For carousel items, this may consist of \code{\link{carouselSets}}s,
which in turn includes one or more \code{\link{carouselItem}}s.}

\item{title}{Optional title.}

\item{footer}{Optional footer text.}

\item{status}{The status of the item This determines the item's background
color. Valid statuses are listed in \link{validStatuses}.}

\item{solidHeader}{Should the header be shown with a solid color background?}

\item{background}{If NULL (the default), the background of the box will be
white. Otherwise, a color string. Valid colors are listed in
\link{validColors}.}

\item{width}{The width of the box, using the Bootstrap grid system. This is
used for row-based layouts. The overall width of a region is 12, so the
default valueBox width of 4 occupies 1/3 of that width. For column-based
layouts, use \code{NULL} for the width; the width is set by the column that
contains the box.}

\item{height}{The height of a box, in pixels or other CSS unit. By default
the height scales automatically with the content.}

\item{collapsible}{If TRUE, display a button in the upper right that allows
the user to collapse the box.}

\item{collapsed}{If TRUE, start collapsed. This must be used with
\code{collapsible=TRUE}.}

\item{.list}{An optional list containing items to put in the carousel same as
the \code{...} arguments, but in list format. This can be useful when working
with programmatically generated items.}
}
\description{
A \code{carouseleBox} displays \code{carouselSets} and \code{carouselItem} as
main box content, often structures as an \code{carouselSets} follows several
\code{carouselItem}s.
}
\seealso{
\code{\link{box}} for usage examples.

Other boxes: \code{\link{box}}, \code{\link{calendarBox}},
  \code{\link{chatBox}}, \code{\link{infoBox}},
  \code{\link{tabBox}}, \code{\link{timelineBox}},
  \code{\link{valueBox}}
}

