% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxes.R
\name{chatBox}
\alias{chatBox}
\alias{chatContact}
\alias{chatContactList}
\alias{chatMessage}
\title{Create a chat box for the main body of a dashboard.}
\usage{
chatBox(..., textInputId, btnInputId, placeholder = "Type Message ...",
  title = "chatBox", status = "warning", solidHeader = FALSE,
  background = NULL, width = 6, height = NULL, admin = "Guang Yang",
  adminImg = paste0("ygdashboard", "-",
  as.character(utils::packageVersion("ygdashboard")), "/img/yg.jpg"),
  client = "Ex Machina", clientImg = paste0("ygdashboard", "-",
  as.character(utils::packageVersion("ygdashboard")), "/img/exmachina.jpg"),
  contactList = NULL, newMessage = NULL)

chatMessage(name, image, text, position = c("left", "right"),
  timestamp = "Just Now")

chatContactList(..., .list = NULL)

chatContact(name, image = NULL, date = "Just Now", text = NULL)
}
\arguments{
\item{...}{For chat message, this should consist of \code{\link{chatMessage}}s.}

\item{title}{Optional title.}

\item{status}{The status of the item This determines the item's background
color. Valid statuses are listed in \link{validStatuses}.}

\item{solidHeader}{Should the header be shown with a solid color background?}

\item{background}{If NULL (the default), the background of the box will be
white. Otherwise, a color string. Valid colors are listed in
\link{validColors}.}

\item{width}{The width of the box, using the Bootstrap grid system. This is
used for row-based layouts. The overall width of a region is 12, so the
default valueBox width of 4 occupies 1/3 of that width. For column-based
layouts, use \code{NULL} for the width; the width is set by the column that
contains the box.}

\item{height}{The height of a box, in pixels or other CSS unit. By default
the height scales automatically with the content.}

\item{.list}{An optional list containing messages to put in the chatBox same as
the \code{...} arguments, but in list format. This can be useful when working
with programmatically generated chatMessage.}

\item{...}{For chat contacts, this should consist of \code{\link{chatContact}}s.}
}
\description{
A \code{chatBox} displays \code{chatMessage}s as main box content, an optinal
  \code{contactList} on header, and an optional \code{newMessage} reminder on
  header. The \code{contactList} are created by \code{\link{chatContactList}}
  which in turn contains multiple \code{\link{chatContact}}s.

Create a chat contact list for the chat box header.
}
\seealso{
\code{\link{box}} for usage examples.

Other boxes: \code{\link{box}}, \code{\link{calendarBox}},
  \code{\link{carouselBox}}, \code{\link{infoBox}},
  \code{\link{socialWigetBox}}, \code{\link{tabBox}},
  \code{\link{timelineBox}}, \code{\link{valueBox}}
}

