% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxes.R
\name{chatMessageOutput}
\alias{chatMessageOutput}
\title{Create a dynamic chat message output for ygdashboard (client side)}
\usage{
chatMessageOutput(outputId)
}
\arguments{
\item{outputId}{Output variable name.}
}
\description{
This can be used as a placeholder for dynamically-generated \code{\link{chatMessage}}.
}
\seealso{
\code{\link{renderChatMessage}} for the corresponding server side function
  and examples.

Other box outputs: \code{\link{renderChatMessage}}
}

