% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/menuOutput.R
\name{menuSegmentOutput}
\alias{menuSegmentOutput}
\title{Create a sidebar menu segment output (client side)}
\usage{
menuSegmentOutput(outputId)
}
\arguments{
\item{outputId}{Output variable name.}
}
\description{
This is the UI-side function for creating a dynamic sidebar menu segment.
}
\seealso{
\code{\link{renderMenu}} for the corresponding server-side function
  and examples, and \code{\link{menuSegment}} for the corresponding function
  for generating static sidebar menu segments.

Other menu outputs: \code{\link{dropdownMenuOutput}},
  \code{\link{menuItemOutput}}, \code{\link{menuOutput}},
  \code{\link{renderMenu}},
  \code{\link{sidebarMenuOutput}},
  \code{\link{sidebarUserPanelOutput}}
}

