% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboardFooter.R
\name{renderFooter}
\alias{renderFooter}
\title{Create dynamic footer (server side)}
\usage{
renderFooter(expr, env = parent.frame(), quoted = FALSE, func = NULL)
}
\arguments{
\item{expr}{An expression that returns a Shiny tag object, \code{\link{HTML}},
or a list of such objects.}

\item{env}{The environment in which to evaluate \code{expr}.}

\item{quoted}{Is \code{expr} a quoted expression (with \code{quote()})? This
is useful if you want to save an expression in a variable.}

\item{func}{A function that returns a Shiny tag object, \code{\link{HTML}},
or a list of such objects (deprecated; use \code{expr} instead).}
}
\description{
Create dynamic footer (server side)
}
\seealso{
\code{\link{footerOutput}} for the corresponding client side function
  and examples.

Other footer outputs: \code{\link{footerOutput}}
}

