% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/menuOutput.R
\name{sidebarUserPanelOutput}
\alias{sidebarUserPanelOutput}
\title{Create a sidebar user panel output (client side)}
\usage{
sidebarUserPanelOutput(outputId)
}
\arguments{
\item{outputId}{Output variable name.}
}
\description{
This is the UI-side function for creating a dynamic sidebar user panel.
}
\seealso{
\code{\link{renderMenu}} for the corresponding server-side function
  and examples, and \code{\link{sidebarUserPanel}} for the corresponding function
  for generating static sidebar menus.

Other menu outputs: \code{\link{dropdownMenuOutput}},
  \code{\link{menuItemOutput}}, \code{\link{menuOutput}},
  \code{\link{menuSegmentOutput}},
  \code{\link{renderMenu}}, \code{\link{sidebarMenuOutput}}
}

