% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxes.R
\name{socialWigetBox}
\alias{socialWigetBox}
\title{Create a socialWidgetBox box for the main body of a dashboard.}
\usage{
socialWigetBox(..., name = NULL, image = NULL, description = NULL,
  stat1 = NULL, subStat1 = NULL, stat2 = NULL, subStat2 = NULL,
  stat3 = NULL, subStat3 = NULL, background = NULL,
  backgroundImage = NULL, width = 6, height = NULL, collapsible = FALSE,
  collapsed = FALSE)
}
\arguments{
\item{...}{Contents of the box.}

\item{background}{If NULL (the default), the background of the box will be
white. Otherwise, a color string. Valid colors are listed in
\link{validColors}.}

\item{width}{The width of the box, using the Bootstrap grid system. This is
used for row-based layouts. The overall width of a region is 12, so the
default valueBox width of 4 occupies 1/3 of that width. For column-based
layouts, use \code{NULL} for the width; the width is set by the column that
contains the box.}

\item{height}{The height of a box, in pixels or other CSS unit. By default
the height scales automatically with the content.}

\item{collapsible}{If TRUE, display a button in the upper right that allows
the user to collapse the box.}

\item{collapsed}{If TRUE, start collapsed. This must be used with
\code{collapsible=TRUE}.}
}
\description{
A \code{socialWidgetBox} as a specialized box that can be used to hold
content in the main body of a dashboard.
}
\seealso{
\code{\link{box}} for usage examples.

Other boxes: \code{\link{box}}, \code{\link{calendarBox}},
  \code{\link{carouselBox}}, \code{\link{chatBox}},
  \code{\link{infoBox}}, \code{\link{tabBox}},
  \code{\link{timelineBox}}, \code{\link{valueBox}}
}

