% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxes.R
\name{timelineBox}
\alias{timelineBox}
\alias{timelineItem}
\alias{timelineLabel}
\title{Create a timeline box for the main body of a dashboard.}
\usage{
timelineBox(..., .list = NULL)

timelineLabel(text, color)

timelineItem(icon = shiny::icon("bars bg-blue"), header = NULL,
  body = NULL, footer = NULL, itemIcon = shiny::icon("clock-o"),
  itemText = "Just Now")
}
\arguments{
\item{...}{For timeline items, this may consist of \code{\link{timelineLabel}}s
and \code{\link{timelineItem}}s.}

\item{.list}{An optional list containing items to put in the timeline same as
the \code{...} arguments, but in list format. This can be useful when working
with programmatically generated items.}

\item{footer}{Optional footer text.}
}
\description{
A \code{timelineBox} displays \code{timelineLabel} and \code{timelineItem} as
main box content, often structures as an \code{timelineLabel} follows several
\code{timelineItem}s.
}
\seealso{
\code{\link{box}} for usage examples.

Other boxes: \code{\link{box}}, \code{\link{calendarBox}},
  \code{\link{carouselBox}}, \code{\link{chatBox}},
  \code{\link{infoBox}}, \code{\link{socialWigetBox}},
  \code{\link{tabBox}}, \code{\link{valueBox}}
}

